+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Documentation improvements
+
+* Some typo fixes and wordings improved
+
 
 ## [1.0.0-alpha.0] - 2023-05-20